Fix circular dependency with page node utils
