+++ src/utils/nodes/page/page.ts
@@ -4,12 +4,11 @@
  * @description Utility functions for page nodes in the editor.
  */
 
-import { Node as PMNode, ResolvedPos } from "@tiptap/pm/model";
+import { Node as PMNode } from "@tiptap/pm/model";
 import { EditorState } from "@tiptap/pm/state";
 import { PAGE_NODE_NAME } from "../../../constants/page";
 import { NodePosArray } from "../../../types/node";
 import { Nullable } from "../../../types/record";
-import { isPosMatchingEndOfPageCondition, isPosMatchingStartOfPageCondition } from "../../pagination";
 
 /**
  * Check if the given node is a page node.
@@ -60,44 +59,4 @@ export const collectPageNodes = (doc: PMNode): NodePosArray => {
     });
 
     return pageNodes;
-};
-
-/**
- * Check if the given position is at the start of the page or the first child of the page.
- * @param doc - The document node.
- * @param $pos - The resolved position in the document or the absolute position of the node.
- * @returns {boolean} True if the condition is met, false otherwise.
- */
-export const isPosAtStartOfPage = (doc: PMNode, $pos: ResolvedPos | number): boolean => {
-    return isPosMatchingStartOfPageCondition(doc, $pos, true);
-};
-
-/**
- * Check if the given position is at the first paragraph child of the page.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {boolean} True if the position is at the start of the page, false otherwise.
- */
-export const isPosAtFirstChildOfPage = (doc: PMNode, $pos: ResolvedPos | number): boolean => {
-    return isPosMatchingStartOfPageCondition(doc, $pos, false);
-};
-
-/**
- * Check if the given position is exactly at the end of the page.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {boolean} True if the position is at the end of the page, false otherwise.
- */
-export const isPosAtEndOfPage = (doc: PMNode, $pos: ResolvedPos | number): boolean => {
-    return isPosMatchingEndOfPageCondition(doc, $pos, true);
-};
-
-/**
- * Check if the given position is at the last paragraph child of the page.
- * @param doc - The document node.
- * @param pos - The resolved position in the document or the absolute position of the node.
- * @returns {boolean} True if the position is at the end of the page, false otherwise.
- */
-export const isPosAtLastChildOfPage = (doc: PMNode, $pos: ResolvedPos | number): boolean => {
-    return isPosMatchingEndOfPageCondition(doc, $pos, false);
 };
